backend/pyproject.toml
[project]
name = "context-graph"
version = "0.1.0"
description = "Context Graph Demo - Decision traces for AI agents using Neo4j"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "neo4j>=5.26.0",
    "neo4j-agent-memory[openai]>=0.0.2",
    "claude-agent-sdk>=0.1.18",
    "openai>=1.58.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "faker>=33.0.0",
    "httpx>=0.28.0",
    "sse-starlette>=2.0.0",
    "graphdatascience>=1.19",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
